+++ wp-content/plugins/shola-core/shola-core.php
@@ -49,6 +49,7 @@ spl_autoload_register( 'shcore_autoload' );
 
 \SholaCore\Post_Types::init();
 \SholaCore\Taxonomies::init();
+\SholaCore\Meta_Fields::init();
 
 /**
  * On activation: register post types and taxonomies, seed the fixed term